+++ P9/Models/SolicitudPrestamo.cs
@@ -6,7 +6,7 @@ namespace P9.AutoModel
   public partial class SolicitudPrestamo
   {
     public int Id { get; set; }
-    public int UsuarioId { get; set; }
+    public long UsuarioId { get; set; }
     public int PrestamoId { get; set; }
     public int Estatus { get; set; }
